pyproject.toml
[tool.poetry]
authors = ["tomoya-n <s266227@wakayama-u.ac.jp>"]
description = ""
name = "pm25"
packages = [{include = "*", from = "src"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
astor = "^0.8.1"
black = "^24.4.2"
codetokenizer = {path = "/Users/tomoya-n/dev/PM25/dist/codetokenizer-0.1.0.tar.gz"}
dataclasses-json = "^0.6.7"
flake8 = "^7.1.0"
gitpython = "^3.1.43"
isort = "^5.13.2"
japanize-matplotlib = "^1.1.3"
matplotlib = "^3.9.0"
nltk = "^3.8.1"
numpy = "^1.26.4"
polars = "^1.4.0"
pytest = "^8.2.2"
python = ">=3.12,<3.14"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
seaborn = "^0.13.2"
setuptools = "^75.8.0"
sklearn-model = "^0.0.6"
tokenizers = "^0.15.2"
tqdm = "^4.66.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.isort]
line_length = 119
profile = "black"

[tool.pytest.ini_options]
testpaths = "test"
